// packages/tentacles/Services/Interfaces/node_web_interface/src/client/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type HTTPValidationError = {
    detail?: Array<ValidationError>;
};

export type Node = {
    node_type: string;
    backend_type: string;
    workers: (number | null);
    status: string;
    redis_url?: (string | null);
    sqlite_file?: (string | null);
};

export type Task = {
    id?: string;
    name?: (string | null);
    description?: (string | null);
    content?: (string | null);
    content_metadata?: (string | null);
    type?: (string | null);
    status?: (TaskStatus | null);
    result?: (string | null);
    result_metadata?: (string | null);
    retries?: (number | null);
    retry_delay?: (number | null);
    priority?: (number | null);
    expires?: (string | null);
    expires_resolved?: (string | null);
    scheduled_at?: (string | null);
    started_at?: (string | null);
    completed_at?: (string | null);
};

export type TaskStatus = 'pending' | 'scheduled' | 'periodic' | 'running' | 'completed' | 'failed';

export type User = {
    email: string;
    is_active?: boolean;
    is_superuser?: boolean;
    full_name?: (string | null);
    id: string;
};

export type ValidationError = {
    loc: Array<(string | number)>;
    msg: string;
    type: string;
};

export type ServeSpaAppRootResponse = (unknown);

export type ServeSpaAppData = {
    path: string;
};

export type ServeSpaAppResponse = (unknown);

export type LoginTestAuthResponse = (User);

export type NodesGetCurrentNodeResponse = (Node);

export type TasksCreateTasksData = {
    requestBody: Array<Task>;
};

export type TasksCreateTasksResponse = ([
    number,
    number
]);

export type TasksGetTasksData = {
    limit?: number;
    page?: number;
};

export type TasksGetTasksResponse = (Array<Task>);

export type TasksUpdateTaskData = {
    requestBody: Task;
    taskId: string;
};

export type TasksUpdateTaskResponse = (Task);

export type TasksDeleteTaskData = {
    taskId: string;
};

export type TasksDeleteTaskResponse = (string);

export type TasksGetMetricsResponse = (unknown);

export type UsersReadUserMeResponse = (User);